+++ addons/binding/org.openhab.binding.icloud/src/main/java/org/openhab/binding/icloud/handler/ICloudAccountBridgeHandler.java
@@ -28,10 +28,10 @@ import org.eclipse.smarthome.core.thing.ThingStatusDetail;
 import org.eclipse.smarthome.core.thing.binding.BaseBridgeHandler;
 import org.eclipse.smarthome.core.thing.binding.ThingHandler;
 import org.eclipse.smarthome.core.types.Command;
-import org.openhab.binding.icloud.discovery.DeviceDiscovery;
 import org.openhab.binding.icloud.internal.Connection;
 import org.openhab.binding.icloud.internal.DeviceInformationParser;
 import org.openhab.binding.icloud.internal.configuration.AccountThingConfiguration;
+import org.openhab.binding.icloud.internal.discovery.DeviceDiscovery;
 import org.openhab.binding.icloud.internal.json.icloud.DeviceInformation;
 import org.openhab.binding.icloud.internal.json.icloud.JSONRootObject;
 import org.osgi.framework.ServiceRegistration;
