+++ addons/binding/org.openhab.binding.icloud/src/main/java/org/openhab/binding/icloud/internal/ICloudHandlerFactory.java
@@ -22,9 +22,9 @@ import org.eclipse.smarthome.core.thing.ThingUID;
 import org.eclipse.smarthome.core.thing.binding.BaseThingHandlerFactory;
 import org.eclipse.smarthome.core.thing.binding.ThingHandler;
 import org.eclipse.smarthome.core.thing.binding.ThingHandlerFactory;
-import org.openhab.binding.icloud.discovery.DeviceDiscovery;
 import org.openhab.binding.icloud.handler.ICloudAccountBridgeHandler;
 import org.openhab.binding.icloud.handler.ICloudDeviceHandler;
+import org.openhab.binding.icloud.internal.discovery.DeviceDiscovery;
 import org.osgi.framework.ServiceRegistration;
 import org.osgi.service.component.annotations.Component;
